requirements.txt
boto3
pyaudio
asyncio
websockets
aiofile
amazon_transcribe
streamlit
uvicorn
flask
